[host][abigen] Add a missing cast
Stricter warnings for host code complain about implicit size_t->int.

Test: manual

+++ system/host/abigen/header_generator.cpp
@@ -16,7 +16,7 @@ static const string add_attribute(map<string, string> attributes,
 
 static vector<int> collect_nonnull(const Syscall& sc) {
     vector<int> nonnull;
-    int out_idx = sc.arg_spec.size();
+    int out_idx = static_cast<int>(sc.arg_spec.size());
     sc.for_each_return([&](const TypeSpec& type) {
             ++out_idx;
             if (!has_attribute("optional", type.attributes))
